Remove assets subdirectory
