+++ input/aws/aws.go
@@ -19,12 +19,14 @@ package aws
 
 import (
 	"context"
-	"github.com/aws/aws-sdk-go-v2/service/ec2"
-	stateless "github.com/elastic/beats/v7/filebeat/input/v2/input-stateless"
 	"time"
 
+	"github.com/aws/aws-sdk-go-v2/service/ec2"
+
+	stateless "github.com/elastic/beats/v7/filebeat/input/v2/input-stateless"
+
 	input "github.com/elastic/beats/v7/filebeat/input/v2"
-	"github.com/elastic/inputrunner/input/assets/internal"
+	"github.com/elastic/inputrunner/input/internal"
 
 	"github.com/elastic/beats/v7/libbeat/feature"
 	conf "github.com/elastic/elastic-agent-libs/config"
