+++ input/aws/ec2.go
@@ -22,7 +22,7 @@ import (
 	"fmt"
 
 	stateless "github.com/elastic/beats/v7/filebeat/input/v2/input-stateless"
-	"github.com/elastic/inputrunner/input/assets/internal"
+	"github.com/elastic/inputrunner/input/internal"
 	"github.com/elastic/inputrunner/util"
 
 	"github.com/elastic/elastic-agent-libs/logp"
