+++ input/aws/ec2_test.go
