+++ input/aws/vpc_test.go
