+++ input/gcp/asset.go
@@ -19,7 +19,7 @@ package gcp
 
 import (
 	"github.com/elastic/elastic-agent-libs/mapstr"
-	"github.com/elastic/inputrunner/input/assets/internal"
+	"github.com/elastic/inputrunner/input/internal"
 )
 
 func WithAssetLabels(value mapstr.M) internal.AssetOption {
