+++ input/gcp/gcp.go
@@ -19,13 +19,14 @@ package gcp
 
 import (
 	"context"
-	"google.golang.org/api/option"
 	"time"
+
+	"google.golang.org/api/option"
 
 	input "github.com/elastic/beats/v7/filebeat/input/v2"
 	stateless "github.com/elastic/beats/v7/filebeat/input/v2/input-stateless"
 	"github.com/elastic/go-concert/ctxtool"
-	"github.com/elastic/inputrunner/input/assets/internal"
+	"github.com/elastic/inputrunner/input/internal"
 
 	"github.com/elastic/beats/v7/libbeat/feature"
 	conf "github.com/elastic/elastic-agent-libs/config"
