+++ input/gcp/gcp_test.go
