+++ input/gcp/gke_test.go
