+++ input/gcp/shared.go
