+++ input/gcp/shared_test.go
