+++ input/internal/config.go
