+++ input/internal/ecs.go
