+++ input/internal/publish.go
