+++ input/internal/publish_test.go
