+++ input/k8s/k8s.go
@@ -25,7 +25,7 @@ import (
 
 	input "github.com/elastic/beats/v7/filebeat/input/v2"
 	stateless "github.com/elastic/beats/v7/filebeat/input/v2/input-stateless"
-	"github.com/elastic/inputrunner/input/assets/internal"
+	"github.com/elastic/inputrunner/input/internal"
 
 	kube "github.com/elastic/elastic-agent-autodiscover/kubernetes"
 	conf "github.com/elastic/elastic-agent-libs/config"
