+++ tests/e2e/assets_aws_test.go
@@ -29,10 +29,11 @@ import (
 	stateless "github.com/elastic/beats/v7/filebeat/input/v2/input-stateless"
 	"github.com/elastic/inputrunner/input/testutil"
 
+	"github.com/stretchr/testify/assert"
+
 	"github.com/elastic/elastic-agent-libs/config"
 	"github.com/elastic/elastic-agent-libs/logp"
-	"github.com/elastic/inputrunner/input/assets/aws"
-	"github.com/stretchr/testify/assert"
+	"github.com/elastic/inputrunner/input/aws"
 )
 
 func TestAssetsAWS_Run_startsAndStopsTheInput(t *testing.T) {
